+++ Interface/AddOns/NDui/Locales/deDE.lua
@@ -916,3 +916,5 @@ L["ChargesRemaining"] = "%s %s/%s next charge remaining %s"
 L["ChargesCompleted"] = "%s %s/%s all charges ready."
 L["ToggleActionbarTip"] = "|nOption to toggle actionbar, no need to reload UI."
 L["IgnoreNoteHelp"] = "You can add note for ignore list by double click." -- need translation
+L["RaidTargetX"] = "RaidTarget X Offset" -- need translation
+L["RaidTargetY"] = "RaidTarget Y Offset" -- need translation