+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # NDui
 
-## [7.9.2](https://github.com/siweia/NDui/tree/7.9.2) (2023-5-27)
+## [7.10.0](https://github.com/siweia/NDui/tree/7.10.0) (2023-6-24)
 
 - Updates for 10.1.5.
+- Update default skins.
+- Update buff frame timer.
 - Update auras and nameplate filter.
+- Fix chat frame quick scrolling.
+- New tier charges on infobar.
+- Update friends infobar game icon texture.
+- Updates for PremadeGroupsFilter 4.0.
+- Option to modify nameplate raid target indicator.